refactor(kolme-runtime): extract runtime error module file (#1960)
